add SGLD Optimizer
Since the SGLD step multiplies the loss gradient by a factor of N
we also use a learning rate which is proportional to (1/N)^2. Accordingly,
we change the lr parameter into lr_factor parameter. Thus, learning rate
will be lr_factor*(1/N)^2, where N is the size of the training dataset

Signed-off-by: Guy Heller <hellerguyh@gmail.com>