attackv02-part1: save predictions and accuracy along the epochs
We want to get the privacy loss along the training epochs. To that end, we need
the predictions over the malicious image along the epochs. We save them in a
file called meta_<model_id>

While here, we also save in meta file meta-data about the run, including
accuracy and loss along the epochs.

Signed-off-by: Guy Heller <hellerguyh@gmail.com>